cart functionality add

+++ controllers/productController.js
@@ -12,43 +12,52 @@ const addProduct = (req, res) => {
     description,
     inCart,
     inWish,
+    tags,
   } = req.body;
 
   const { email } = req.user;
   User.findOne({ email })
     .then((response) => {
-      cloudinary.v2.uploader.upload(
-        req.file.path,
-        { public_id: "ecommerce-app/products/" + req.file.filename },
-        function (err, result) {
-          if (err) {
-            serverError(res);
-          } else {
-            const product = {
-              author: {
-                authorId: response._id,
-                email: response.email,
-              },
-              name,
-              category,
-              regularPrice,
-              salePrice,
-              description,
-              inCart,
-              inWish,
-              image: result.url,
-            };
-            new Product(product)
-              .save()
-              .then((response) => {
-                res.status(200).json(response);
-              })
-              .catch(() => {
-                serverError(res);
-              });
+      if (response.role === "admin" || response.role === "vendor") {
+        cloudinary.v2.uploader.upload(
+          req.file.path,
+          { public_id: "ecommerce-app/products/" + req.file.filename },
+          function (err, result) {
+            if (err) {
+              serverError(res);
+            } else {
+              const product = {
+                author: {
+                  authorId: response._id,
+                  storeName: response.storeName,
+                },
+                name,
+                category,
+                regularPrice,
+                salePrice,
+                description,
+                tags: JSON.parse(tags),
+                inCart,
+                inWish,
+                image: [
+                  {
+                    url: result.url,
+                    public_id: result.public_id,
+                  },
+                ],
+              };
+              new Product(product)
+                .save()
+                .then((response) => {
+                  res.status(200).json(response);
+                })
+                .catch(() => {
+                  serverError(res);
+                });
+            }
           }
-        }
-      );
+        );
+      }
     })
     .catch(() => {
       serverError(res);
@@ -75,21 +84,31 @@ const updateProduct = (req, res) => {
     description,
     inCart,
     inWish,
+    tags,
   } = req.body;
 
-  const updateProduct = {
-    name,
-    category,
-    regularPrice,
-    salePrice,
-    description,
-    inCart,
-    inWish,
-  };
-
-  Product.findOneAndUpdate({ _id: id, updateProduct })
+  const { email } = req.user;
+  User.findOne({ email })
     .then((response) => {
-      res.status(200).json(response);
+      if (response.role === "admin" || response.role === "vendor") {
+        const updateProduct = {
+          name,
+          category,
+          regularPrice,
+          salePrice,
+          description,
+          tags: JSON.parse(tags),
+          inCart,
+          inWish,
+        };
+        Product.findOneAndUpdate({ _id: id, updateProduct }, { new: true })
+          .then((response) => {
+            res.status(200).json(response);
+          })
+          .catch(() => {
+            serverError(res);
+          });
+      }
     })
     .catch(() => {
       serverError(res);
@@ -98,9 +117,18 @@ const updateProduct = (req, res) => {
 
 const deleteProduct = (req, res) => {
   const id = req.params.id;
-  Product.findOneAndRemove({ _id: id })
+  const { email } = req.user;
+  User.findOne({ email })
     .then((response) => {
-      res.status(200).json(response);
+      if (response.role === "admin" || response.role === "vendor") {
+        Product.findOneAndRemove({ _id: id })
+          .then((response) => {
+            res.status(200).json(response);
+          })
+          .catch(() => {
+            serverError(res);
+          });
+      }
     })
     .catch(() => {
       serverError(res);
